refactor: (tobe) 상품 패키지 구조 리팩터링
