+++ src/main/java/kitchenpos/products/tobe/domain/JpaProductRepository.java
@@ -1,7 +1,8 @@
-package kitchenpos.products.domain.tobe.domain;
+package kitchenpos.products.tobe.domain;
 
 import kitchenpos.common.domain.ProductId;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, ProductId> {
+
 }
