+++ src/main/java/kitchenpos/products/tobe/domain/ProductRepository.java
@@ -1,10 +1,11 @@
-package kitchenpos.products.domain.tobe.domain;
+package kitchenpos.products.tobe.domain;
 
 import java.util.List;
 import java.util.Optional;
 import kitchenpos.common.domain.ProductId;
 
 public interface ProductRepository {
+
     Product save(Product product);
 
     Optional<Product> findById(ProductId id);
