+++ src/test/java/kitchenpos/menus/tobe/domain/menu/MenuValidatorTest.java
@@ -14,9 +14,9 @@ import kitchenpos.common.domain.MenuGroupId;
 import kitchenpos.common.domain.ProductId;
 import kitchenpos.menus.tobe.application.InMemoryMenuGroupRepository;
 import kitchenpos.menus.tobe.domain.menugroup.MenuGroupRepository;
-import kitchenpos.products.application.tobe.InMemoryProductRepository;
-import kitchenpos.products.domain.tobe.domain.Product;
-import kitchenpos.products.domain.tobe.domain.ProductRepository;
+import kitchenpos.products.tobe.application.InMemoryProductRepository;
+import kitchenpos.products.tobe.domain.Product;
+import kitchenpos.products.tobe.domain.ProductRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
